build a grid with the size of user input
there are also some other shape functions you can call.  Which is what I was doing when I first started writing the program.  Then, I thought it would be interesting if I could create a grid with the size being whatever I wanted it to be at run-time.  This was the result of my messing around.